requirements.txt
flask
tensorflow
pillow
groq
reportlab
requests